+++ packages/implementations/react/src/atomic/Skeleton/src/skeleton.types.ts
@@ -1,0 +1,11 @@
+import { HTMLAttributes } from "react";
+import { SkeletonSprinkle } from "@rarui/styles";
+
+export interface SkeletonProperties extends SkeletonSprinkle {
+  /**
+   * This is an attribute used to identify a DOM node for testing purposes.
+   */
+  "data-testid"?: string;
+}
+
+export type SkeletonProps = SkeletonProperties & HTMLAttributes<HTMLDivElement>;