+++ packages/styles/src/index.ts
@@ -7,11 +7,13 @@ export { box } from "./packages/atomic/box";
 export { button } from "./packages/atomic/button";
 export { chip } from "./packages/atomic/chip";
 export { link } from "./packages/atomic/link";
+export { skeleton } from "./packages/atomic/skeleton";
 export { text } from "./packages/atomic/text";
 export { title } from "./packages/atomic/title";
 
 // types
 export type { BoxSprinkle } from "./packages/atomic/box";
 export type { TextSprinkle } from "./packages/atomic/text";
+export type { SkeletonSprinkle } from "./packages/atomic/skeleton";
 export type { TitleSprinkle } from "./packages/atomic/title";
 export type { ButtonVariants } from "./packages/atomic/button";